Add a StatusItem (MenuExtra) as UI
Add a StatusMenuItem as UI

We now also keep track of which chains should be running, so we can
re-enable them after sleep / wake (or relaunch).

We also listen for device notifications, so we remove yanked chains.

Pretty much all that's missing is an icon (both for menu and app).

Well, that and testing ^_^