Compute require min input frames at output time
Instead of trying to be clever and adjusting (and getting it wrong when
the input cycle size is not a divisor of the framesPerOutputPacket),
we assume a constant input cycle size, and use that to compute how
many input frames we need to keep around.

Also, instead of outputting silence when missing a packet, we repeat
the last packet.