Reduce latency by setting IOCycleUsage
By default, CoreAudio assumes we need 100% CPU for filling the output
buffer. Even for encoding this is not the case, so we can tell it to
schedule us later for reduced latency.

On my Mac Pro, I need ~0.01 of the output cycle. By test encoding
over a number of trials, taking the minimum and then multiplying that
by a large safety factor (cache effects etc.), I estimate  using 0.02
parts of the cycle, which gives us a reduction in latency of about
1500 frames (~0.03s at 48kHz).