Reduce buffer (cycle) size for forwarding
The default buffer size that CoreAudio chooses (barring other
considerations such as device restrictions etc.) is 512 frames.

This means that in the worst case, the delay before an input signal
arrives at the output IOProc is that many frames (and about half as
much on average). By explicitly requesting a smaller size (128), we
reduce the average latency for that part of processing from ~256
to ~64 frames (about 0.004s at 48kHz).