Support FFmpeg 5.0
This meant having to find a lot of workarounds for heap operations in
the happy path (mainly coming from the reference-counted AVBuffers which
forcibly allocate new AVBufferRefs all the time) when using the
recommended public methods for encoding.

Instead of calling the (now removed) avcodec_encode_audio2() we now call
the audio codec's encode2 directly.

A similar approach is done for the muxer (SPDIF) to avoid it otherwise
making a ton of calls to `free(nullptr)` when cleaning up non-existent
side-data in a temporary packet.