Move to CATaps for obtaining input
Instead of requiring an audio device that mirrors the audio data it
receives back to its input (looking like a microphone), we now use a
CoreAudioTap on a private aggregate device to listen in on the output
it receives.

That lets us work with just "system audio recording" permissions. These
can be found in:
System Settings -> Privacy & Security -> Screen & System Audio Recording

Microphone access is no longer needed *unless* the output device that
is tapped also has inputs (which is currently still the case for
SoundPusherAudio).