SoundPusherAudio: Remove loopback (and inputs)
This removes the loopback functionality from SoundPusherAudio; it now
simply swallows all audio and does nothing else.

The output mute control was also removed (since it is now pointless, and
all it did was mute its mirrored input).

This also means we no longer have a use for CARingBuffer and therefore
can remove it.